module4-task1 вынес допиливание данных в стор
